# ⚡ Improve performance for claiming rewards