--- a/src/Application/Common/Constants/CacheKeys.cs
+++ b/src/Application/Common/Constants/CacheKeys.cs
@@ -4,6 +4,7 @@ public static class CacheKeys
 {
     public const string Leaderboard = "LeaderboardAllTime";
     public const string UserRanking = "UserRankingsAllTime";
+    public const string ClaimPrizeAchievementId = "ClaimPrizeAchievementId";
 }
 
 /// <summary>