> 1. What triggered this change? (PBI link, Email Subject, conversation + reason, etc)

✏️ #1160 

> 2. What was changed?

✏️ Improved DB indexes and queries to fetch data for claiming rewards.

Original SQL query for the top-most active user returned 1000+ rows with ~42 columns! Now it will always return only 1 row.

The execution plan difference was done based on my account, which only had ~100 rows in the original query; however, the analysis shows very promising improvements. The query was handcrafted and ChatGPT o3 was used to verify the execution plan.

## ChatGPT summary of original vs new SQL query

| Metric | Original plan – fan-out joins | After indexes – scalar sub-queries |
| --- | --- | --- |
| Rows returned to client | 93 detail rows (User × Rewards × Achievements) | 1 row (totals + flag)
| Sub-tree cost | 0.0652 | 0.0301  (- 54 %)
| Actual CPU / elapsed | ≈ 1 ms / 1 ms | 0 ms / < 1 ms
| Logical reads (total) | ≈ 147 pages | ≈ 91 pages  (- 38 %)
| - Users | 4 | 4
| - UserRewards | 6 (clustered scan) | 2 (seek on IX_UserRewards_UserId)
| - Rewards | 18 | 16 (seek on IX_Rewards_Id)
| - UserAchievements | 67 | 3
| - Achievements | 62 (PK look-ups) | 64 (seek on IX_Achievements_Id)
| Memory grant | 1 024 KB granted (16 KB used) | 0 KB
| Compile memory / CPU | 1 568 KB / 10 ms | 1 152 KB / 5 ms
| Network payload | ≈ 340 KB | ≈ 8 KB
| Largest scan | 529-row PK scan of UserRewards | none; all seeks, max 64 look-ups
| Look-ups / rewinds | 62 PK look-ups on Achievements, Lazy-Spool rewinds | 32 Achievements look-ups, 8 Rewards look-ups, no spool
| Wait types / spills | none | none

### Summary

After adding the two covering non-clustered indexes the query:

* Eliminates full clustered scans and the spool.
* Cuts logical reads by ≈ 40 %.
* Needs no memory grant and reports zero CPU/elapsed time.

For a single-user dashboard call this is “good enough”; any further tuning would bring negligible benefit until the tables are at least two orders of magnitude larger.

**NOTE: My computer is too fast with too much RAM (64GB) to make any meaningful performance measurements. (~150 ms)  This is why I used execution plans to verify the performance.**

> 3. Did you do pair or mob programming?

✏️ No.
<!-- E.g. I worked with @gordonbeeming and @sethdailyssw -->

<!-- 
Check out the relevant rules
- https://www.ssw.com.au/rules/rules-to-better-pull-requests
- https://www.ssw.com.au/rules/write-a-good-pull-request
- https://www.ssw.com.au/rules/over-the-shoulder-prs 
- https://www.ssw.com.au/rules/do-you-use-co-creation-patterns
-->