@jernejk 2025-07-02
## Copilot Summary

This pull request introduces enhancements to the reward claiming functionality and database optimizations. Key changes include refactoring the `ClaimRewardCommandHandler` to improve performance and logging, adding new indexes to database entities for faster queries, and updating migration files to reflect these changes.

### Reward Claiming Enhancements:
* [`src/Application/Rewards/Commands/ClaimReward/ClaimRewardCommand.cs`](diffhunk://#diff-07a82867922080f8df6d969f7c4a8ecb4391ca786cc42cdf0f51ed72f3da0cc7R24-R85): Refactored `ClaimRewardCommandHandler` to use projection-based queries for fetching user data, added logging for better debugging, and implemented caching for the achievement ID lookup. Introduced a helper method `GetClaimPrizeAchievementId` to retrieve the achievement ID efficiently. [[1]](diffhunk://#diff-07a82867922080f8df6d969f7c4a8ecb4391ca786cc42cdf0f51ed72f3da0cc7R24-R85) [[2]](diffhunk://#diff-07a82867922080f8df6d969f7c4a8ecb4391ca786cc42cdf0f51ed72f3da0cc7L75-R116) [[3]](diffhunk://#diff-07a82867922080f8df6d969f7c4a8ecb4391ca786cc42cdf0f51ed72f3da0cc7R127-R138)

* [`src/Application/Common/Constants/CacheKeys.cs`](diffhunk://#diff-1009ff1a3d5ebc96400ab04110c3b88c8f2e8f9adc7763605ecfa9419cb64314R7): Added a new cache key `ClaimPrizeAchievementId` to support caching of the achievement ID.

### Database Index Optimizations:
* [`src/Infrastructure/Persistence/Configurations/AchievementConfiguration.cs`](diffhunk://#diff-88a4348497b61e9fe85e09e65fd36bb1de8b15c90f31b146359a073f9fe0d023R13-R16): Added an index on `Id` with included properties for `Value` to optimize queries involving achievements.

* [`src/Infrastructure/Persistence/Configurations/RewardConfiguration.cs`](diffhunk://#diff-c71e2ea64a7ef89929b7dd4ade44cb9eba9cf85614f8abe80d1517df6d070ee6R11-R13): Added an index on `Id` with included properties for `Cost` to improve reward-related queries.

* [`src/Infrastructure/Persistence/Configurations/UserRewardConfiguration.cs`](diffhunk://#diff-fdcf7dbeab3f55d0b72b5ac37357b8bfb0da44ad3b1e36cf144c58a0bf37afd7R19-R22): Added an index on `UserId` with included properties for `RewardId` to optimize user reward lookups.

### Migration Updates:
* [`src/Infrastructure/Persistence/Migrations/20250702075823_ClaimRewardIndexes.cs`](diffhunk://#diff-1bab4fac7c2976fd856d81945e06e35b26fd3a4e0ee1e020177e6726c04be3c6R1-R57): Updated migration file to include new indexes for `Achievements`, `Rewards`, and `UserRewards`. Dropped and recreated the `UserRewards` index to include `RewardId`.

### Model Snapshot Updates:
* [`src/Infrastructure/Persistence/Migrations/ApplicationDbContextModelSnapshot.cs`](diffhunk://#diff-ec04a38e5cb2c9e35197741d63019149cebdc089133d658f4000ac07ca177ee1L20-R20): Updated the model snapshot to reflect the new indexes added to `Achievements`, `Rewards`, and `UserRewards`. Changed the EF Core product version annotation to `9.0.6`. [[1]](diffhunk://#diff-ec04a38e5cb2c9e35197741d63019149cebdc089133d658f4000ac07ca177ee1L20-R20) [[2]](diffhunk://#diff-ec04a38e5cb2c9e35197741d63019149cebdc089133d658f4000ac07ca177ee1R65-R68) [[3]](diffhunk://#diff-ec04a38e5cb2c9e35197741d63019149cebdc089133d658f4000ac07ca177ee1R450-R453) [[4]](diffhunk://#diff-ec04a38e5cb2c9e35197741d63019149cebdc089133d658f4000ac07ca177ee1R775-R776)